% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\docType{class}
\name{address-class}
\alias{address-class}
\alias{address}
\title{A RC class to create address objects}
\description{
A RC class to create address objects
}
\section{Fields}{

\describe{
\item{\code{address}}{A character address to look for}

\item{\code{latlong}}{A 2-lentghed numeric vector with longitude and latitude}
}}


