% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolada.R
\name{advancedSearch}
\alias{advancedSearch}
\title{Advanced Search (Search by specific KPI code, Municipality code or Year)}
\usage{
advancedSearch(kpi_list = NULL, municipality_list = NULL, year_list = NULL)
}
\arguments{
\item{kpi_list}{Id of the KPIs as a list.}

\item{municipality_list}{Id of the municipalities as a list.}

\item{year_list}{The years to get as a list.}
}
\value{
a list that contains data about KPIs in a given municipality by a specific (user-defined) year.
}
\description{
Advanced Search (Search by specific KPI code, Municipality code or Year)
}
