% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{API}
\alias{API}
\alias{my_API}
\title{Kolada API}
\usage{
my_API(url)
}
\arguments{
\item{url}{data}
}
\description{
functions you should have to access the data in the API
}
\examples{
output <- my_API("http://api.kolada.se/v2/ou?title=skola")
}
\references{
\url{https://www.kolada.se/om-oss/api/}
}
\author{
Mahnaz , Bita
}
