% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Run the Kolada KPI with Shiny app}
\usage{
app(ui, server)
}
\arguments{
\item{ui}{shiny UI}

\item{server}{shiny server}
}
\description{
This function runs the Shiny app that allows us to display the data from 
the Kolada KPI. The app provides an interface so you can select municipality, KPI and year and 
then creates a bar plot divided by gender.
}
