% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5.R
\name{bikeStationStatus}
\alias{bikeStationStatus}
\title{Get information about the busiest and the least busy city bike station in a city}
\usage{
bikeStationStatus(network_ids)
}
\arguments{
\item{city_info}{The network information for a city.}
}
\value{
A list containing the busiest and the least busy city bike station details.
}
\description{
This function takes the network information for a city and returns details about the busiest and the least busiest station of the city bikes in the selected city.
}
