% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ca_2019}
\alias{Ca_2019}
\title{Ca_2019}
\format{A data frame consisting 34 observations on 5 variables.}
\description{
A dataset containingv five variables: longi, latit, GoodDays,
    ModerateDays, UnhealthyDays. This data set shows the number of the days
    in 2019 based on how was the quality of weather.
    
 
\itemize{
  \item{longi:} {Longitude of the locations.}
  \item{latit:} {Latitude of the locations}
  \item{GoodDays:} {Number of healthy days.}
  \item{ModerateDays:} {Number of moderate days.}
  \item{UnhealthyDays:} {Number of unhealthy days.}
}
}
\references{
\url{https://en.wikipedia.org/wiki/Graph_(abstract_data_type)}
}
\keyword{datasets}
