% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada.R
\name{compare_inhabitants}
\alias{compare_inhabitants}
\title{This is a function used to compare the evolution of inhabitation numbers of 
swedish municipalities}
\usage{
compare_inhabitants(municipality1, municipality2, return_data = FALSE)
}
\arguments{
\item{municipality1}{The first municipality to be comapred. 
ATTENTION: Exact spelling including special letters is needed (e.g. ö)}

\item{municipality2}{The second municipality to be comapred
ATTENTION: Exact spelling including special letters is needed (e.g. ö)}

\item{return_data}{Per default this is FALSE, if it is TRUE no plot is 
output, instead a data.frame is returned}
}
\description{
This function takes two names of swedish municipalities as arguments and
returns either a plot of the progression of their inhabitant numbers or
 (if the return_data argument is TRUE) returns a data.frame with the inhabitant
 data of those municipalities. Data usually exists for the period 1970-2021.
}
\examples{
compare_inhabitants("Linköping","Norrköping")
compare_inhabitants("Linköping","Norrköping", return_data = TRUE)


}
