% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epochPOSIX.R
\name{epochPOSIX}
\alias{epochPOSIX}
\title{Converts UNIX epoch time to CEST time}
\usage{
epochPOSIX(gbgjsondate)
}
\arguments{
\item{gbgjsondate}{The JSON date format used by th city of Gothenburg open
  data API}
}
\value{
raw A POSIX string showing the time in CEST, DST adjusted.
}
\description{
Takes JSON date character string and returns POSIX character string, CEST
time zone DST adjusted
}
\details{
This function takes a string of form \code{"/Date(1234567891000+XX00)/"} and
returns a string containing the corresponding POSIX, CEST time zone time.
}
\examples{
epochPOSIX("XXXXXX123465786456578")
}

