% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kolada_data.R
\name{fetch_all_given_entity}
\alias{fetch_all_given_entity}
\title{Retrieving data for given kpi, municipality and years}
\usage{
fetch_all_given_entity(kpi, municipality_id, year)
}
\arguments{
\item{kpi}{Entity KPI}

\item{municipality_id}{Passing the id of municipality}

\item{year}{A vector of years}
}
\value{
Returns a data.frame containing the data for a given kpi, municipality and year
}
\description{
Retrieving data for given kpi, municipality and years
}
\examples{
fetch_all_given_entity("N00945",1860,c(2009,2007,2008))
}
