% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kolada_data.R
\name{fetch_given_kpiandmuncipality_id}
\alias{fetch_given_kpiandmuncipality_id}
\title{Retrieving data for given kpi and municipality.}
\usage{
fetch_given_kpiandmuncipality_id(kpi, municipality_id)
}
\arguments{
\item{kpi}{Entity KPI}

\item{municipality_id}{Passing the id of municipality}
}
\value{
Returns a data.frame containing the data for a given kpi and municipality. Data displayed are Count : number of embers associated with the selected KPIs and Municipality ID, Gender (Man , Kvinna and total),status : if the average is available and Value : weighted average
}
\description{
Retrieving data for given kpi and municipality.
}
\examples{
fetch_given_kpiandmuncipality_id("N00945",1860)
}
