% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kolada_data.R
\name{fetch_with_given_kpiandyear}
\alias{fetch_with_given_kpiandyear}
\title{Retrieving data for given kpi and years}
\usage{
fetch_with_given_kpiandyear(kpi, year)
}
\arguments{
\item{kpi}{Entity KPI}

\item{year}{A vector of years}
}
\value{
Returns a data.frame containing the data for a given kpi and year
}
\description{
Retrieving data for given kpi and years
}
\examples{
fetch_with_given_kpiandyear("N00905",2009)
}
