% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kolada_data.R
\name{fetch_with_given_muncipalityandyear}
\alias{fetch_with_given_muncipalityandyear}
\title{Retrieving data for given municipality and years}
\usage{
fetch_with_given_muncipalityandyear(municipality_id, year)
}
\arguments{
\item{municipality_id}{Passing the id of municipality}

\item{year}{A vector of years}
}
\value{
Returns a data.frame containing the data for a given municipality and year
}
\description{
Retrieving data for given municipality and years
}
\examples{
fetch_with_given_muncipalityandyear(1860,2009)
}
