% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5.R
\name{fetchCityBikeData}
\alias{fetchCityBikeData}
\title{city_bikes documentation}
\usage{
fetchCityBikeData(
  api_urls = c("http://api.citybik.es/v2/networks/lundahoj",
    "http://api.citybik.es/v2/networks/malmobybike")
)
}
\arguments{
\item{api_urls}{blablabla}
}
\value{
A list containing the busiest and the least busy city bike station details.
}
\description{
taking the information from city bikes API. With the information we get two
cities and find the least busy and busy stations. Show them to the user.
fetchCityBikeData function handling the APIs and fins the busy and least busy station.
}
