% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_laureates.r
\name{filter_laureates}
\alias{filter_laureates}
\title{Filter the laureates' data frame}
\usage{
filter_laureates(
  dframe,
  gender = "",
  awardYear = "",
  awardYearTo = "",
  name = "",
  birthYear = "",
  birthYearTo = "",
  country = "",
  continent = "",
  category = "",
  age = "",
  limit = 2000
)
}
\arguments{
\item{dframe}{data frame}

\item{gender}{"male", "female", "org"}

\item{awardYear}{integer}

\item{awardYearTo}{integer}

\item{name}{string}

\item{birthYear}{integer (YYYY)}

\item{birthYearTo}{integer (YYYY)}

\item{country}{string}

\item{continent}{string ['Africa', 'Asia', 'Europe', 'North America', 'Oceania', 'South America']}

\item{category}{string ['che', 'eco', 'lit', 'pea', 'phy', 'med']}

\item{age}{integer}

\item{limit}{integer}
}
\value{
dframe data frame
}
\description{
This function filters the database of Nobel Laureates.
The input data frame must be the result of parse_to_shiny.
}
