% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5.R
\name{findBusiestStationsFromData}
\alias{findBusiestStationsFromData}
\title{Find the busiest and least busy city bike stations for a given city data.}
\usage{
findBusiestStationsFromData(city_data)
}
\arguments{
\item{city_data}{A JSON data file containing information about the city's bike network.}
}
\value{
A list containing the busiest and the least busy city bike station details.
}
\description{
This function takes city data and returns details about
the busiest and the least busy city bike stations in the selected city.
}
