% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo1.R
\name{geo_connect}
\alias{geo_connect}
\title{Title}
\usage{
geo_connect(address)
}
\arguments{
\item{address}{}
}
\value{
A data.frame containing the resulting \code{longitude/s}, \code{latitude/s} and \code{address/es}
}
\description{
Title
}
\examples{
geo_connect("delhi")
}
