% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo2.R
\name{geo_connect_vector}
\alias{geo_connect_vector}
\title{Title}
\usage{
geo_connect_vector(address)
}
\arguments{
\item{address}{}
}
\value{
A data.frame containing the resulting \code{longitude/s}, \code{latitude/s} and \code{address/es}
}
\description{
Title
}
\examples{
geo_connect_vector(c("40.41678, -3.70379", " Linkoping"))
}
