% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_func.R
\docType{class}
\name{geocode_func-class}
\alias{geocode_func-class}
\alias{geocode_func}
\title{Coordinates finder}
\arguments{
\item{data}{A data frame of character string addreses}
}
\value{
A single coordinate or a dataframe cointaining full address names as well as the coordinates
}
\description{
The 'geocode_func' function is a class that converts an address to coordinates.
}

\examples{
e <- geocode_func()
e$getCoordinates( "Mobile, Alabama, United States")

e$getCorrdinates(addresses)

}
