% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodification.R
\docType{package}
\name{lab5}
\alias{geocodification}
\alias{lab5-package}
\title{A package for connecting to Google Geocoding API}
\description{
The \code{lab5} package provides an interface to connect with the \code{Google Geocoding API} with
functions that convert adresses into coordinates and vice versa.
}
\details{
The package provides the following function:
\itemize{
   \item \code{geocodification}: Google Geocoding API connector
}
}
\author{
\itemize{
     \item Alejandro Garcia - \url{https://github.com/Raikao}
     \item Ramon Laborda - \url{https://github.com/ramonlaborda}
   }
}
\references{
Google Geocoding API - \url{https://developers.google.com/maps/documentation/geocoding/intro}
}

