% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api.R
\name{get_api}
\alias{get_api}
\title{Fetch Data from Kolada API}
\usage{
get_api(search, api_key)
}
\arguments{
\item{search}{A string representing the search term for the API query.}

\item{api_key}{A string representing the API key for authentication.}
}
\value{
Parsed content from the API response.
}
\description{
This function fetches data from the Kolada API based on the search term provided and the user's API key.
}
