% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_channels.R
\name{get_channels}
\alias{get_channels}
\title{Returns a data frame with channels and channels.id}
\usage{
get_channels()
}
\value{
channel_file A dataframe with channel names and channels.id.
}
\description{
This function gets all the channel names and ids from SR v.2. API
}

