% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coordinates.R
\name{get_coordinates}
\alias{get_coordinates}
\title{getLocation}
\usage{
get_coordinates(city_name)
}
\arguments{
\item{city_name}{The name of a city as a string}
}
\value{
The latitude and longitude coordinates of the city
}
\description{
getLocation
}
