% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(kpi, municipality, year)
}
\arguments{
\item{kpi}{id of the kpi to fetch data for}

\item{municipality}{name of the municipality to fetch data for}

\item{year}{characters of the year to fetch data for}
}
\value{
Will return a data frame with fetched data from API for given parameters
}
\description{
yuppy
}
\examples{
get_data("N09890", "Helsingborg", "2019")

}
