% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kolada_data.R
\name{get_kolada_data}
\alias{get_kolada_data}
\title{Get Kolada Data}
\arguments{
\item{year}{The year for which to fetch data.}
}
\value{
A list containing the data for the specified year.
}
\description{
Obtain the data for a specific year.
}
