% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KPI.R
\name{get_kpi}
\alias{get_kpi}
\title{Get list of KPIs from Kolada API}
\usage{
get_kpi()
}
\value{
A data frame of municipalities
}
\description{
Get list of KPIs from Kolada API
}
\examples{
KPIs <- get_kpi()
head(KPIs)

}
