% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kpi_group.R
\name{get_kpi_group}
\alias{get_kpi_group}
\title{Get KPI group}
\usage{
get_kpi_group(group_id)
}
\arguments{
\item{group_id}{id of the kpi group}
}
\value{
Will return a vector containing all kpis for a kpi group
}
\description{
yuppy
}
\examples{
get_kpi_group("GKPI127")

}
