% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipalities.R
\name{get_municipalities}
\alias{get_municipalities}
\title{Get list of municipalities from Kolada API}
\usage{
get_municipalities()
}
\value{
A data frame of municipalities
}
\description{
Get list of municipalities from Kolada API
}
\examples{
municipalities <- get_municipalities()
head(municipalities)

}
