% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OU.R
\name{get_ou}
\alias{get_ou}
\title{Get list of Organization Unit from Kolada API}
\usage{
get_ou()
}
\value{
A data frame of municipalities
}
\description{
Get list of Organization Unit from Kolada API
}
\examples{
organizational_unit <- get_ou()
head(organizational_unit)

}
