% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_programs.R
\name{get_programs}
\alias{get_programs}
\title{Returns a data frame with programs and channels.id for the programs}
\usage{
get_programs(channel_frame)
}
\arguments{
\item{channel_frame}{A dataframe with channel names and channel.id.}
}
\value{
program_file A dataframe with program names and broadcast channel.id
  and channel name.
}
\description{
This function gets all the program names and channel.id and names for the
channels on which the programs are broadcast from SR v.2. API
}
\details{
This function only returns the programs which are not yet archived, i.e.
still broadcasting.(purportedly)
}

