% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_song.R
\name{get_song}
\alias{get_song}
\title{Returns a data frame with The song and artist playing currently}
\usage{
get_song(channel_name, channel_frame)
}
\arguments{
\item{channel_name}{A string containing the channel name.}

\item{channel_frame}{A data frame with channel names and channels.id.}
}
\value{
song_file A dataframe with song name and performing artist.
}
\description{
This function returns the current song and artist on the specified channel
name from SR v.2. API.
}

