% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kolada_data.R
\name{getMetadata}
\alias{getMetadata}
\title{Get an object from the Kolada API}
\usage{
getMetadata(entity, title = NULL)
}
\arguments{
\item{entity}{The entity of search query for KPI, municipality, OU, kpi_groups, municipality_groups}
}
\value{
Returns a data.frame containing the data corresponding to the entity
}
\description{
Get an object from the Kolada API
}
\examples{
getMetadata("municipality","Helsingborg")
}
