% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_api.R
\name{google_api}
\alias{google_api}
\title{Google_API}
\usage{
google_api(address)
}
\description{
google_api is a function to find lat, lon codes and detail address by inputing an address.
}
\examples{
google_api("Eiffel+Tower,Paris")
google_api("Summer+Palace,Beijing")

}

