% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5Code.R
\docType{data}
\name{Iowa}
\alias{Iowa}
\title{ia_results}
\format{A data frame with 100 rows and 3 variables
\describe{
 \item{Vote by county}{County of interest}
 \item{Trump}{Total votes for Donald Trump}
 \item{Clinton}{Total votes for Hillary Clinton}
}}
\source{
\url{http://www.nytimes.com/elections/results/iowa}
}
\usage{
Iowa
}
\description{
Total Votes for Trump and Clinton in Iowa by county
A data set containing the raw vote totals for Donald Trump and Hillary Clinton for each of Iowa's 100 counties.
}
\keyword{datasets}
