% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada.R
\docType{class}
\name{kolada-class}
\alias{kolada-class}
\alias{kolada}
\title{A class to collect data from kolada}
\description{
A class to collect data from kolada
}
\section{Fields}{

\describe{
\item{\code{municipal}}{character: A given municipality}
}}

\section{Methods}{

\describe{
\item{\code{get_id()}}{Initialize the values}

\item{\code{get_school()}}{Get the school list with given municipality}
}}

