% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_1.R
\name{kolada_api}
\alias{kolada_api}
\title{kolada_api}
\usage{
kolada_api(path)
}
\arguments{
\item{path}{character}
}
\value{
a character string
}
\description{
kolada_api() is used for finding more information about certain city.
}
