% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{kpi}
\alias{kpi}
\title{KPIs}
\usage{
kpi(id = NULL, title = NULL, page = 1, per_page = 5000)
}
\arguments{
\item{id}{vector of KPI ids; if given, filters by id}

\item{title}{substring match on title}

\item{page}{Page number}

\item{per_page}{results per page}
}
\value{
Kolada envelope: list(values, count, next_page, previous_page)
}
\description{
KPIs
}
