% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolada.R
\name{Kpis}
\alias{Kpis}
\title{Kpis("kvinnofridskr채nkning") : http://api.kolada.se/v2/kpi?title=kvinnofridskr채nkning}
\usage{
Kpis(name = NULL, BYgroups = FALSE)
}
\arguments{
\item{name}{The name of the kpi}

\item{BYgroups}{For KPI group search}
}
\value{
specific KPI or kPI groups
}
\description{
Kpis("kvinnofridskr채nkning") : http://api.kolada.se/v2/kpi?title=kvinnofridskr채nkning
}
