\name{Lab5-package}
\alias{Lab5-package}
\alias{Lab5}
\docType{package}
\title{
\packageTitle{Lab5}
}
\description{
\packageDescription{Lab5}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab5}
\packageIndices{Lab5}
}
\author{
\packageAuthor{Lab5}

Maintainer: \packageMaintainer{Lab5}
}
\keyword{ package }