% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5-package.R
\docType{package}
\name{lab5}
\alias{lab5}
\title{A package city coordinate visualisation and p-norm distance calculation.}
\description{
This package offers a get_coordinates function returning the coordinates of a input city.
Furthermore a shiny application calculating the p-norm distance between two swedish cities is provided.
}
\details{
lab5
}
