% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5class.R
\docType{class}
\name{lab5class-class}
\alias{lab5class-class}
\alias{lab5class}
\title{Coordinate finder}
\description{
The 'lab5class' using API from thenmap
}
\section{Fields}{

\describe{
\item{\code{res}}{A data.frame including id and name}

\item{\code{name_of_data}}{Name of dataset you want}

\item{\code{availist}}{Vector of available name of name_of_data}
}}


