% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{The initializer function for the data}
\usage{
main()
}
\description{
This function initializes channels and program .csv files for use in Shinyapp
project by executing SR Open API requests.
}

