% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolada.R
\name{Municipalities}
\alias{Municipalities}
\title{Municipalities("Linköping") : http://api.kolada.se/v2/municipality?title=Linköping
Municipalities("0580", BYgroups=TRUE) : http://api.kolada.se/v2/ou?municipality=0580}
\usage{
Municipalities(name = NULL, BYgroups = FALSE)
}
\arguments{
\item{name}{The name of the municipality we want to access}

\item{BYgroups}{For municipality group search}
}
\value{
Specific Municipality or Municipality groups
}
\description{
Municipalities("Linköping") : http://api.kolada.se/v2/municipality?title=Linköping
Municipalities("0580", BYgroups=TRUE) : http://api.kolada.se/v2/ou?municipality=0580
}
