% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{municipality}
\alias{municipality}
\title{Municipality}
\usage{
municipality(id = NULL, title = NULL, type = NULL, page = 1, per_page = 5000)
}
\arguments{
\item{id}{vector of municipalities ids; if given, filters by id}

\item{title}{substring match on title}

\item{type}{substring match on type. It can be "municipality" or "region"}

\item{page}{Page number}

\item{per_page}{results per page}
}
\value{
Kolada envelope: list(values, count, next_page, previous_page)
}
\description{
Municipality
}
