% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolada.R
\name{Municipality_groups}
\alias{Municipality_groups}
\title{Municipality Groups (Get a municipality's organizational units)}
\usage{
Municipality_groups(name)
}
\arguments{
\item{name}{The name/code id of the municipality's organizational.}
}
\value{
list of all organizational units of a municipality.
}
\description{
Municipality Groups (Get a municipality's organizational units)
}
