% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_to_shiny.r
\name{parse_to_shiny}
\alias{parse_to_shiny}
\title{Modify laureate database for shiny app}
\usage{
parse_to_shiny(dframe)
}
\arguments{
\item{dframe}{data frame}
}
\value{
dframe data frame
}
\description{
This function selects some columns from the laureates' database
and adds others that are needed in the shiny app.
The columns added are age, awardYear, category and prizeAmount.
}
