% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_over_time.R
\name{plot_age_over_time}
\alias{plot_age_over_time}
\title{Plot the age of laureates at the time of their award}
\usage{
plot_age_over_time(laureates)
}
\arguments{
\item{laureates}{A data frame returned by \code{\link[=get_laureates]{get_laureates()}}.}
}
\value{
A ggplot object: one point per laureate with a trend line,
}
\description{
Plot the age of laureates at the time of their award
}
