% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_map.R
\name{plot_birth_countries}
\alias{plot_birth_countries}
\title{Map the birth countries of laureates}
\usage{
plot_birth_countries(laureates)
}
\arguments{
\item{laureates}{A data frame returned by \code{\link[=get_laureates]{get_laureates()}}.}
}
\value{
A ggplot object: a world map coloured by number of laureates born
in each country.
}
\description{
Map the birth countries of laureates
}
