% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sf_data.R
\name{plot_sf_data}
\alias{plot_sf_data}
\title{Plot geo-data}
\usage{
plot_sf_data(sf_data)
}
\arguments{
\item{sf_data}{geo-dataset in sf format}
}
\value{
The function returns a geo-plot
}
\description{
This function can be used plot geo-data
}
\examples{
library(Lab5)
requested_data <- request_API("v1","world-2","geo","2016-01-01","fi")
plot_sf_data(requested_data)

}
\seealso{
\url{https://www.thenmap.net/}
}
