% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5Code.R
\name{proportion}
\alias{proportion}
\title{Proportion}
\usage{
proportion(Vec1, Vec2, byrow = TRUE)
}
\arguments{
\item{Vec1}{a Vector}

\item{Vec2}{a Vector}

\item{byrow}{logical.  If \code{TRUE} (default), the proportions will be calculated by row, otherwise by column}
}
\value{
The proportion of the value in the first row or column specified by the user
}
\description{
This is a function that will take the proportions of an element of one vector comparing to the similar element of another vector
}
\examples{
proportion(c(1:10),c(11:20))
proportion(c(1:100),c(101:200))
}
\author{
Ashish Jain, Matt Stuart, Haiyang Zhang
}
