% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{q}
\alias{q}
\title{Functions.}
\format{\preformatted{List of 10
 $ url        : chr "http://www.val.se/val/val2014/statistik/2014_riksdagsval_per_kommun.xls"
 $ status_code: int 200
 $ headers    :List of 9
  ..$ date          : chr "Wed, 21 Oct 2015 08:34:00 GMT"
  ..$ server        : chr "Apache"
  ..$ last-modified : chr "Fri, 03 Oct 2014 06:18:22 GMT"
  ..$ etag          : chr "\\"10c358-1d400-5047eb32cc780\\""
  ..$ accept-ranges : chr "bytes"
  ..$ content-length: chr "119808"
  ..$ cache-control : chr "max-age=3600"
  ..$ expires       : chr "Wed, 21 Oct 2015 09:34:00 GMT"
  ..$ content-type  : chr "application/vnd.ms-excel"
  ..- attr(*, "class")= chr [1:2] "insensitive" "list"
 $ all_headers:List of 1
  ..$ :List of 3
  .. ..$ status : int 200
  .. ..$ version: chr "HTTP/1.1"
  .. ..$ headers:List of 9
  .. .. ..$ date          : chr "Wed, 21 Oct 2015 08:34:00 GMT"
  .. .. ..$ server        : chr "Apache"
  .. .. ..$ last-modified : chr "Fri, 03 Oct 2014 06:18:22 GMT"
  .. .. ..$ etag          : chr "\\"10c358-1d400-5047eb32cc780\\""
  .. .. ..$ accept-ranges : chr "bytes"
  .. .. ..$ content-length: chr "119808"
  .. .. ..$ cache-control : chr "max-age=3600"
  .. .. ..$ expires       : chr "Wed, 21 Oct 2015 09:34:00 GMT"
  .. .. ..$ content-type  : chr "application/vnd.ms-excel"
  .. .. ..- attr(*, "class")= chr [1:2] "insensitive" "list"
 $ cookies    :'data.frame':	0 obs. of  7 variables:
  ..$ domain    : logi(0) 
  ..$ flag      : logi(0) 
  ..$ path      : logi(0) 
  ..$ secure    : logi(0) 
  ..$ expiration:Classes 'POSIXct', 'POSIXt'  num(0) 
  ..$ name      : logi(0) 
  ..$ value     : logi(0) 
 $ content    : raw [1:119808] d0 cf 11 e0 ...
 $ date       : POSIXct[1:1], format: "2015-10-21 08:34:00"
 $ times      : Named num [1:6] 0 0.00493 0.00514 0.00516 0.10835 ...
  ..- attr(*, "names")= chr [1:6] "redirect" "namelookup" "connect" "pretransfer" ...
 $ request    :List of 7
  ..$ method    : chr "GET"
  ..$ url       : chr "http://www.val.se/val/val2014/statistik/2014_riksdagsval_per_kommun.xls"
  ..$ headers   : Named chr "application/json, text/xml, application/xml, */*"
  .. ..- attr(*, "names")= chr "Accept"
  ..$ fields    : NULL
  ..$ options   :List of 2
  .. ..$ useragent    : chr "libcurl/7.43.0 r-curl/0.9.3 httr/1.0.0"
  .. ..$ customrequest: chr "GET"
  ..$ auth_token: NULL
  ..$ output    : list()
  .. ..- attr(*, "class")= chr [1:2] "write_memory" "write_function"
  ..- attr(*, "class")= chr "request"
 $ handle     :Class 'curl_handle' <externalptr> 
 - attr(*, "class")= chr "response"
}}
\usage{
q
}
\arguments{
\item{No}{params}
}
\description{
Uses a GET-verb to connect to val.se and downloads a xls-file, which is
converted and read by a read.xlsx-command into the actual data-file we need. From this
file it extracts the things we need. The shiny app then reads this by calling the
package Lab5, which only consists of this data.
}
\author{
Maxime Bonneau, Niclas Lovsjö
}
\keyword{datasets}

