% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uclassifier.R
\name{remove_classifier}
\alias{remove_classifier}
\title{Delete Classifier}
\usage{
remove_classifier(classifier)
}
\arguments{
\item{classifier}{The UClassifier to delete.}
}
\description{
Function for delete your classifier
}

