% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_API.R
\name{request_API}
\alias{request_API}
\title{Fetch data from Thenmaps}
\usage{
request_API(dataset, date)
}
\arguments{
\item{dataset}{The dataset is a area-level. For the area, the ISO 3166 codes where available was used, and for level, the ten levels of administrative borders from Open Streetmap was used.}

\item{date}{Can be on the form YYYY-MM-DD, YYYY-MM, or YYYY. The output will include only areas that existed at that very date. Dates in the database are local time and inclusive, meaning that 1947-03-01 will include a nation that gained independence on April 1 1947, according to the local timezone. 2011-05 will be interpreted as 2001-05-01, and 2011 will be interpreted as 2001-01-01.}
}
\value{
The function returns the requested dataset.
}
\description{
This function can be used to fetch data from the Thenmaps-API.
}
\examples{
library(Lab5)
requested_data <- request_API("world-2", "2016-01-01")

}
\seealso{
\url{https://www.thenmap.net/}
}
