% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolada.R
\name{respond}
\alias{respond}
\title{Check Server Response and fetch data}
\usage{
respond(api_call)
}
\arguments{
\item{api_call}{API's from server response}
}
\value{
API values
}
\description{
Check Server Response and fetch data
}
