% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riksdagen.R
\name{riksdagen}
\alias{riksdagen}
\title{THis is a description of riksdagen function}
\usage{
riksdagen(voteyear)
}
\arguments{
\item{voteyear}{a character with fixed format "YYYY/YY"}
}
\value{
a data frame with percentage of voting data
}
\description{
riksdagen function uses Riksdagen's API to get data and makes some basic analysis.
}
\details{
The function uses the API to get the voting history grouped by party and makes some basic analysis.
}
\examples{
 riksdagen("2020/21")

}
