% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRiksShiny.R
\name{runRiksShiny}
\alias{runRiksShiny}
\title{runRiksShiny}
\usage{
runRiksShiny()
}
\value{
runs a Shiny app locally
}
\description{
runRiksShiny runs an included shiny app allowing the user to see which way each party votes in total by year
}
\examples{
 \dontrun{runRiksShiny()}

}
