% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Runs the shiny app to demonstrate the compare_inhabitants function visually}
\usage{
runShiny()
}
\description{
Runs the shiny app to demonstrate the compare_inhabitants function visually
}
