\name{shinyapp-package}
\alias{shinyapp-package}
\alias{shinyapp}
\docType{package}
\title{
\packageTitle{shinyapp}
}
\description{
\packageDescription{shinyapp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{shinyapp}
\packageIndices{shinyapp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{shinyapp}

Maintainer: \packageMaintainer{shinyapp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
