% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{shinyapp}
\alias{shinyapp}
\title{Google Geocoding API connector aplication}
\arguments{
\item{address}{A character vector with an address or some coordinates (latitude and longitude)}
}
\value{
A shiny app
}
\description{
Passes the function parameter \code{address} as a parameter to the API via URL.
Then it returns the output of the request into a formatted dataframe with the geocoding information.
}
\references{
Google Geocoding API  - \url{https://developers.google.com/maps/documentation/geocoding/intro}
}
