% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{shinyappserver}
\alias{shinyappserver}
\title{Google Geocoding API connector app}
\arguments{
\item{address}{or some coordinates (latitude and longitude)}
}
\value{
A shiny app containing the resulting \code{longitude/s}, \code{latitude/s} and \code{address/es}
}
\description{
Loops over the vector of addresses and/or coordinates, passes each one of them as a parameter
to an internal function that connects to the \code{Google Geocoding API} via URL and gets the output. Then it
returns the output of the request into a formatted dataframe with the geocoding information.
}
\references{
Google Geocoding API - \url{https://developers.google.com/maps/documentation/geocoding/intro}
}
