% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweden_border.R
\docType{data}
\name{sweden_border}
\alias{sweden_border}
\title{The coordinates of the border of Sweden}
\format{
A dataframe containing two columns (longitude and latitude) and 407398 rows representing points on the border of Sweden.
}
\source{
\url{https://gadm.org/download_country_v3.html}
}
\usage{
sweden_border
}
\description{
This dataset contains a dataframe with the longitude and latitude coordinates of the border of Sweden.
}
\details{
sweden_border

column names: longitude, latitude
}
\keyword{datasets}
