% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{visualize_kolada}
\alias{visualize_kolada}
\title{Visualize Kolada
Starts Kolada visualization demo.}
\format{
An object of class \code{shiny.appobj} of length 5.
}
\usage{
visualize_kolada
}
\value{
Nothing.
}
\description{
Visualize Kolada
Starts Kolada visualization demo.
}
\keyword{datasets}
