% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\docType{class}
\name{address-class}
\alias{address-class}
\alias{address}
\title{A RC class to create address objects}
\description{
A RC class to create address objects
}
\details{
Please look at the vignette.
}
\section{Fields}{

\describe{
\item{\code{address_or_coordinates}}{A character address to look for OR A 2-lentghed numeric vector with longitude and latitude}
}}


\examples{
my_address<-address$new("Linkoping university")
}
