% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_API.R
\name{earthquake_API}
\alias{earthquake_API}
\title{Access the USGS Eartchquake API}
\usage{
earthquake_API(
  format = "csv",
  starttime = "2026-10-01",
  endtime = "2026-10-05",
  min_magnitude = -1,
  max_magnitude = 15
)
}
\arguments{
\item{format}{CSV DO NOT CHANGE}

\item{starttime}{Limit to events on or after the specified end start Enter character string "YYYY-MM-DD"}

\item{endtime}{Limit to events on or before the specified end time. Enter character string "YYYY-MM-DD"}

\item{min_magnitude}{Limit to events with a magnitude larger than the specified minimum, numeric}

\item{max_magnitude}{Limit to events with a magnitude smaller than the specified maximum, numeric}
}
\value{
a data.frame containing time, latitude, longitude, magnitude and location name of events
}
\description{
Access the USGS Eartchquake API
}
