% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_API.R
\name{earthquake_API}
\alias{earthquake_API}
\title{Access the USGS Eartchquake API}
\usage{
earthquake_API(
  format = "csv",
  starttime = "2026-10-01",
  endtime = "2026-10-05",
  min_magnitude = -1,
  max_magnitude = 15
)
}
\arguments{
\item{starttime}{start date- character string format YYYY-MM-DD}

\item{endtime}{end date- character string format YYYY-MM-DD}

\item{min_magnitude}{Limit to events with a magnitude larger than the specified minimum}

\item{max_magnitude}{Limit to events with a magnitude smaller than the specified maximum}
}
\value{
a data.frame containing time, latitude, longitude, magnitude and location name of events
}
\description{
Access the USGS Eartchquake API
}
