% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elect_viz.R
\docType{class}
\name{elect_viz-class}
\alias{elect_viz-class}
\alias{elect_viz}
\title{An RC class that manages all the data}
\description{
An RC class that manages all the data
}
\section{Fields}{

\describe{
\item{\code{data}}{The input for the visualization in the format of 2014_riksdagsval_per_*.xls}
}}

\section{Methods}{

\describe{
\item{\code{get_counties()}}{Returns a list of all the available muniplicities}

\item{\code{get_county()}}{Returns the current county}

\item{\code{get_mean_p_vals()}}{Returns the mean p-values for the currently selected county}

\item{\code{initialize(data)}}{Initializes the data set, sets the available counties and
then selects the first county in the list of counties as the default one}

\item{\code{set_county(val)}}{Sets the currently selected county}
}}

