% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo1.R
\name{geo_connect}
\alias{geo_connect}
\title{Google's geocoding API}
\usage{
geo_connect(address)
}
\arguments{
\item{address}{It takes the address of location.}
}
\value{
A data.frame containing the resulting \code{longitude/s}, \code{latitude/s} and \code{address/es}
}
\description{
This funcion returns the coordinates of API.
}
\examples{
geo_connect("delhi")
}
