% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_func.R
\docType{class}
\name{geocode_func-class}
\alias{geocode_func-class}
\alias{geocode_func}
\title{Coordinates finder}
\description{
The 'geocode_func' function is a class that converts an address to coordinates.
}
\section{Fields}{

\describe{
\item{\code{F}}{A data frame}
}}


