% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download data from the Kolada API}
\usage{
get_data(kpi, municipality, year)
}
\arguments{
\item{kpi}{The specific question/query ID in the API}

\item{municipality}{The municipality in the API}

\item{year}{The year in the API}
}
\value{
a data frame of the specific arguments you have selected in the
  function and if no data is available for the given parameters, it returns
  `-1`.
}
\description{
This function retrieves data for a specific KPI for a given municipality and
year from the Kolada API
}
\examples{
# Fetch KPI data from the municipality  Helsingborg for the year 2019
get_data("N09890", "Helsingborg", "2019")
}
