% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(kpi, municipality, year)
}
\value{
Will return a data frame with fetched data from API for given parameters
}
\description{
yuppy
}
\examples{
get_data()

}
