% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kpi_group.R
\name{get_kpi_group}
\alias{get_kpi_group}
\title{Get KPIs for the specific KPI gruop}
\usage{
get_kpi_group(group_id)
}
\arguments{
\item{group_id}{id of the KPI group selected}
}
\value{
A data frame containing all KPIs for the specified KPI group
}
\description{
This function fetches all KPIs for the specific KPI group you have selected
and returns the data for the group ID
}
\examples{
# Get all KPIs for the KPI group GKPI127
get_kpi_group("GKPI127")

}
