% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_laureates.r
\name{get_laureates}
\alias{get_laureates}
\title{Get list of laureates using API calls}
\usage{
get_laureates(
  gender = "",
  year = "",
  yearTo = "",
  name = "",
  birthDate = "",
  birthDateTo = "",
  birthCountry = "",
  category = "",
  limit = 2000
)
}
\arguments{
\item{gender}{"male", "female", "other"}

\item{year}{integer}

\item{yearTo}{integer}

\item{name}{string}

\item{birthDate}{date format (YYYY, -MM-, -DD or YYYY-MM-DD)}

\item{birthDateTo}{date format (YYYY, -MM-, -DD or YYYY-MM-DD)}

\item{birthCountry}{string}

\item{category}{string ['che', 'eco', 'lit', 'pea', 'phy', 'med']}

\item{limit}{integer}
}
\value{
dframe data frame
}
\description{
This function downloads the database of Nobel Laureates from nobelprize.org.
Several filters can be applied on the query to reduce the database size.
}
\references{
https://www.nobelprize.org/about/developer-zone-2/
}
