% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipalities.R
\name{get_municipalities}
\alias{get_municipalities}
\title{Get municipalities}
\usage{
get_municipalities()
}
\value{
Will return a vector containing all municipalities that are a kommun
}
\description{
yuppy
}
\examples{
get_municipalities()

}
