% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipalities.R
\name{get_municipalities}
\alias{get_municipalities}
\title{Get all municipalities from the Kolada API}
\usage{
get_municipalities()
}
\value{
Return a vector containing all municipalities that are categorized
  as a kommun and extract the regions
}
\description{
This function fetches all the municipalities from the Kolada API
  and filter out the regions in the API
}
\examples{
# Get a list of all municipalities (kommuner) from the Kolada API
get_municipalities()

}
