% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5Code.R
\docType{data}
\name{Iowa}
\alias{Iowa}
\title{Iowa}
\format{A data frame with 99 rows and 2 variables
\describe{
 \item{Trump}{Total votes for Donald Trump}
 \item{Clinton}{Total votes for Hillary Clinton}
}}
\source{
\url{http://www.nytimes.com/elections/results/iowa}
}
\usage{
Iowa
}
\description{
Total Votes for Trump and Clinton in Iowa by county
A data set containing the raw vote totals for Donald Trump and Hillary Clinton for each of Iowa's 99 counties.
}
\examples{
row.names(Iowa)
sum(Iowa[,1])
}
\keyword{datasets}
