% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koladaRFC.R
\docType{class}
\name{kolada-class}
\alias{kolada-class}
\alias{kolada}
\title{Kolada}
\description{
Kolada
}
\section{Fields}{

\describe{
\item{\code{result}}{list. Final result}

\item{\code{root_url}}{character. The base url of the API}

\item{\code{municipality}}{numeric. Municipality}

\item{\code{year}}{vector. Year}

\item{\code{kpi}}{list. key performance indicator (KPI)}

\item{\code{kpis}}{fetched kpis from server}
}}


