% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada.R
\docType{class}
\name{kolada-class}
\alias{kolada-class}
\alias{kolada}
\title{A class to collect data from kolada}
\description{
A class to collect data from kolada
}
\section{Methods}{

\describe{
\item{\code{get_id(Mname)}}{Get the id of a municipality}

\item{\code{get_municipality_list()}}{Get the complete municipality list}

\item{\code{get_numb(Mname, op)}}{Get the statistics}

\item{\code{get_skola(municipal)}}{Get the school list in a municipality}
}}

