% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolada.R
\name{kpi_groups}
\alias{kpi_groups}
\title{Get KPI groups}
\usage{
kpi_groups(name)
}
\arguments{
\item{name}{The group's name.}
}
\value{
information about a specific kpi group as a list.
}
\description{
Get KPI groups
}
