% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5-package.R
\docType{package}
\name{Lab5-package}
\alias{Lab5}
\alias{Lab5-package}
\title{Lab5: Shiny web application}
\description{
This package is using Kolada API in JSON format, the package httr in used for the HTTP communication and jsonlite for mapping JSON data to data.frames.
}
\author{
\strong{Maintainer}: Aydin Ardalan \email{aydin.ardalan@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
