\name{lab5-package}
\alias{lab5-package}
\alias{lab5}
\docType{package}
\title{
\packageTitle{lab5}
}
\description{
\packageDescription{lab5}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab4new}
\packageIndices{lab4new}
This Package contains th implementation of google geocode API. Geocoding takes an address and return a coordinate in space. This coordinate, along with other coordinates, can then be meaningfully used and analysed. The coordinate system are longitude (x) and latitude (y) coordinates expressed in decimal degrees and formated address. 
}
\author{ packageAuthor {Syeda Farha Shazmeen}

Maintainer packageMaintainer{Farhana Chowdhury Tondra}
}

\references{
https://rstudio-pubs-static.s3.amazonaws.com/265821_3855c8a18af14214b52fbf67d02a52b3.html
https://gist.github.com/josecarlosgonz/6417633
}

\keyword{ package }
\seealso{
http://r-pkgs.had.co.nz/vignettes.html.
http://r-pkgs.had.co.nz/vignettes.html#markdown.
}
\examples{
geo_connect("Delhi")

}


