\name{lab5-package}
\alias{lab5-package}
\alias{lab5}
\docType{package}
\title{
\packageTitle{lab5}
}
\description{
\packageDescription{lab5}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab5}
\packageIndices{lab5}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{lab5}

Maintainer: \packageMaintainer{lab5}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
