% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_api.R
\docType{package}
\name{Lab5}
\alias{Lab5}
\alias{Lab5-package}
\title{A package connecting to a web API.}
\description{
A package is created which is connected to Google Distance Matrix API,
it gives the distance between two cities and time taken to travel.
}
\section{func1() function}{

func1() returns the distance and time taken to travel between two cities.
}

\examples{
func1("chennai","bangalore")
}
\references{
\href{https://developers.google.com/maps/documentation/javascript/distancematrix}
}
