% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Lab5}
\alias{Lab5}
\alias{Lab5-package}
\title{UClassify}
\description{
A wrapper around the rest-API provided by uclassify.com that makes it possible to create classifiers for text data.
}
\author{
Rasmus Holm

Zhendong Wang

Karolina Ziomek
}

