% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\docType{package}
\name{lab5}
\alias{lab5}
\alias{lab5-package}
\title{Data Collector}
\description{
Implements an interactive tool for the Swedish election results where you can
see the results for each muniplicity. The default select is Stockholms Kommun.
Uses the elect_viz package
}
