\name{lab5ab-package}
\alias{lab5ab-package}
\alias{lab5ab}
\docType{package}
\title{RC class for Google Geooding API}
\description{
\packageDescription{lab5ab}
}
\details{ Please look at the vignette


}
\author{
\packageAuthor{lab5ab}

Maintainer: \packageMaintainer{lab5ab}
}
\keyword{ package }
\examples{
my_address<-address$new("linkoping university")
}
