\name{lab5ab-package}
\alias{lab5ab-package}
\alias{lab5ab}
\docType{package}
\title{
\packageTitle{lab5ab}
}
\description{
\packageDescription{lab5ab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab5ab}
\packageIndices{lab5ab}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lab5ab}

Maintainer: \packageMaintainer{lab5ab}
}
\keyword{ package }
\examples{
}
