% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolada.R
\name{Municipality}
\alias{Municipality}
\title{Municipality}
\usage{
Municipality(name)
}
\arguments{
\item{name}{The municipality's name we want to access.}
}
\value{
list of all municipalities.
}
\description{
Municipality
}
