% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab5Code.R
\name{proportion}
\alias{proportion}
\title{Proportion}
\usage{
proportion(Vec1, Vec2)
}
\arguments{
\item{Vec1}{a numeric vector}

\item{Vec2}{a numeric vector}
}
\value{
The proportion of the values given by the user.
}
\description{
This is a function that will take the proportions of an element of one vector comparing to the similar element of another vector.
}
\examples{
proportion(c(1:10),c(11:20))
proportion(c(1:100),c(101:200))
}
\author{
Ashish Jain, Matt Stuart, Haiyang Zhang
}
