% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uclassifier.R
\docType{class}
\name{UClassifier-class}
\alias{UClassifier}
\alias{UClassifier-class}
\title{UClassifier}
\source{
\url{https://uclassify.com/docs}
}
\description{
UClassifier is a text classifier that uses the rest-API provided by uclassify.com
}
\section{Fields}{

\describe{
\item{\code{classifier_name}}{The name of the classifier.}

\item{\code{username}}{Your username on the website.}

\item{\code{read_token}}{Your read token to access the API.}

\item{\code{write_token}}{Your write token to access the API.}
}}

